좋아요 카운트 변경되게 수정

+++ raccoon-api/src/main/java/kwu/raccoonapi/dto/story/like/response/StoryLikeToggleResponse.java
@@ -8,10 +8,10 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor(staticName = "of")
 @NoArgsConstructor()
 public class StoryLikeToggleResponse {
-    private Boolean likeStatus; // 해당 유저가 좋아요를 눌렀는지 여부
-    private Integer likeCount; // 해당 게시물에 눌린 총 좋아요 수
+    private Boolean likeStatus;
+    private Long likeCount;
 
-    public static StoryLikeToggleResponse of(Integer likeCount, Boolean likeStatus){
+    public static StoryLikeToggleResponse of(Long likeCount, Boolean likeStatus){
         StoryLikeToggleResponse storyLikeToggleResponse = new StoryLikeToggleResponse();
         storyLikeToggleResponse.likeCount= likeCount;
         storyLikeToggleResponse.likeStatus=likeStatus;
